Obtiene datos de mindicador.cl

Se conecta a la API y realiza algunas transformaciones al input en formato JSON para facilitar el uso de estos datos con distintas funciones de R

mindicador_importar_datos(
  series = "uf",
  anios = 2020,
  tipo = "data.frame",
  max_intentos = 5,
  usar_cache = FALSE,
  archivo = NULL
)

Arguments

series
Cualquiera de los codigos de mindicador::mindicador_indicadores

anios
Anios validos para las distintas series (ver mindicador::mindicador_indicadores)

tipo
data.frame o xts

max_intentos
Por defecto son cinco

usar_cache
Valor logico para leer y escribir en cache. Si es TRUE, los resultados quedan cacheados en memoria si archivo es NULL o en disco si archivo es una cadena de texto.

archivo
Ruta del archivo para leer y escribir la cache.

Value

un objeto data.frame o xts con los datos solicitados por el usuario

Examples

mindicador_importar_datos("uf", 2020)

#>     serie      fecha    valor
#> 191    uf 2020-01-01 28310.86
#> 190    uf 2020-01-02 28311.77
#> 189    uf 2020-01-03 28312.68
#> 188    uf 2020-01-04 28313.60
#> 187    uf 2020-01-05 28314.51
#> 186    uf 2020-01-06 28315.42
#> 185    uf 2020-01-07 28316.33
#> 184    uf 2020-01-08 28317.25
#> 183    uf 2020-01-09 28318.16
#> 182    uf 2020-01-10 28319.07
#> 181    uf 2020-01-11 28319.99
#> 180    uf 2020-01-12 28320.90
#> 179    uf 2020-01-13 28321.81
#> 178    uf 2020-01-14 28322.73
#> 177    uf 2020-01-15 28323.64
#> 176    uf 2020-01-16 28324.55
#> 175    uf 2020-01-17 28325.47
#> 174    uf 2020-01-18 28326.38
#> 173    uf 2020-01-19 28327.29
#> 172    uf 2020-01-20 28328.21
#> 171    uf 2020-01-21 28329.12
#> 170    uf 2020-01-22 28330.03
#> 169    uf 2020-01-23 28330.95
#> 168    uf 2020-01-24 28331.86
#> 167    uf 2020-01-25 28332.77
#> 166    uf 2020-01-26 28333.69
#> 165    uf 2020-01-27 28334.60
#> 164    uf 2020-01-28 28335.51
#> 163    uf 2020-01-29 28336.43
#> 162    uf 2020-01-30 28337.34
#> 161    uf 2020-01-31 28338.25
#> 160    uf 2020-02-01 28339.17
#> 159    uf 2020-02-02 28340.08
#> 158    uf 2020-02-03 28341.00
#> 157    uf 2020-02-04 28341.91
#> 156    uf 2020-02-05 28342.82
#> 155    uf 2020-02-06 28343.74
#> 154    uf 2020-02-07 28344.65
#> 153    uf 2020-02-08 28345.56
#> 152    uf 2020-02-09 28346.48
#> 151    uf 2020-02-10 28352.33
#> 150    uf 2020-02-11 28358.18
#> 149    uf 2020-02-12 28364.03
#> 148    uf 2020-02-13 28369.88
#> 147    uf 2020-02-14 28375.73
#> 146    uf 2020-02-15 28381.59
#> 145    uf 2020-02-16 28387.44
#> 144    uf 2020-02-17 28393.30
#> 143    uf 2020-02-18 28399.15
#> 142    uf 2020-02-19 28405.01
#> 141    uf 2020-02-20 28410.87
#> 140    uf 2020-02-21 28416.73
#> 139    uf 2020-02-22 28422.60
#> 138    uf 2020-02-23 28428.46
#> 137    uf 2020-02-24 28434.32
#> 136    uf 2020-02-25 28440.19
#> 135    uf 2020-02-26 28446.06
#> 134    uf 2020-02-27 28451.93
#> 133    uf 2020-02-28 28457.80
#> 132    uf 2020-02-29 28463.67
#> 131    uf 2020-03-01 28469.54
#> 130    uf 2020-03-02 28475.41
#> 129    uf 2020-03-03 28481.29
#> 128    uf 2020-03-04 28487.16
#> 127    uf 2020-03-05 28493.04
#> 126    uf 2020-03-06 28498.92
#> 125    uf 2020-03-07 28504.80
#> 124    uf 2020-03-08 28510.68
#> 123    uf 2020-03-09 28516.56
#> 122    uf 2020-03-10 28520.23
#> 121    uf 2020-03-11 28523.91
#> 120    uf 2020-03-12 28527.58
#> 119    uf 2020-03-13 28531.25
#> 118    uf 2020-03-14 28534.93
#> 117    uf 2020-03-15 28538.60
#> 116    uf 2020-03-16 28542.28
#> 115    uf 2020-03-17 28545.95
#> 114    uf 2020-03-18 28549.63
#> 113    uf 2020-03-19 28553.31
#> 112    uf 2020-03-20 28556.98
#> 111    uf 2020-03-21 28560.66
#> 110    uf 2020-03-22 28564.34
#> 109    uf 2020-03-23 28568.02
#> 108    uf 2020-03-24 28571.70
#> 107    uf 2020-03-25 28575.38
#> 106    uf 2020-03-26 28579.06
#> 105    uf 2020-03-27 28582.74
#> 104    uf 2020-03-28 28586.42
#> 103    uf 2020-03-29 28590.10
#> 102    uf 2020-03-30 28593.78
#> 101    uf 2020-03-31 28597.46
#> 100    uf 2020-04-01 28601.15
#> 99     uf 2020-04-02 28604.83
#> 98     uf 2020-04-03 28608.51
#> 97     uf 2020-04-04 28612.20
#> 96     uf 2020-04-05 28615.88
#> 95     uf 2020-04-06 28619.57
#> 94     uf 2020-04-07 28623.25
#> 93     uf 2020-04-08 28626.94
#> 92     uf 2020-04-09 28630.63
#> 91     uf 2020-04-10 28633.49
#> 90     uf 2020-04-11 28636.35
#> 89     uf 2020-04-12 28639.21
#> 88     uf 2020-04-13 28642.07
#> 87     uf 2020-04-14 28644.93
#> 86     uf 2020-04-15 28647.79
#> 85     uf 2020-04-16 28650.65
#> 84     uf 2020-04-17 28653.51
#> 83     uf 2020-04-18 28656.37
#> 82     uf 2020-04-19 28659.23
#> 81     uf 2020-04-20 28662.09
#> 80     uf 2020-04-21 28664.96
#> 79     uf 2020-04-22 28667.82
#> 78     uf 2020-04-23 28670.68
#> 77     uf 2020-04-24 28673.54
#> 76     uf 2020-04-25 28676.41
#> 75     uf 2020-04-26 28679.27
#> 74     uf 2020-04-27 28682.13
#> 73     uf 2020-04-28 28685.00
#> 72     uf 2020-04-29 28687.86
#> 71     uf 2020-04-30 28690.73
#> 70     uf 2020-05-01 28693.59
#> 69     uf 2020-05-02 28696.46
#> 68     uf 2020-05-03 28699.32
#> 67     uf 2020-05-04 28702.19
#> 66     uf 2020-05-05 28705.05
#> 65     uf 2020-05-06 28707.92
#> 64     uf 2020-05-07 28710.79
#> 63     uf 2020-05-08 28713.65
#> 62     uf 2020-05-09 28716.52
#> 61     uf 2020-05-10 28716.52
#> 60     uf 2020-05-11 28716.52
#> 59     uf 2020-05-12 28716.52
#> 58     uf 2020-05-13 28716.52
#> 57     uf 2020-05-14 28716.52
#> 56     uf 2020-05-15 28716.52
#> 55     uf 2020-05-16 28716.52
#> 54     uf 2020-05-17 28716.52
#> 53     uf 2020-05-18 28716.52
#> 52     uf 2020-05-19 28716.52
#> 51     uf 2020-05-20 28716.52
#> 50     uf 2020-05-21 28716.52
#> 49     uf 2020-05-22 28716.52
#> 48     uf 2020-05-23 28716.52
#> 47     uf 2020-05-24 28716.52
#> 46     uf 2020-05-25 28716.52
#> 45     uf 2020-05-26 28716.52
#> 44     uf 2020-05-27 28716.52
#> 43     uf 2020-05-28 28716.52
#> 42     uf 2020-05-29 28716.52
#> 41     uf 2020-05-30 28716.52
#> 40     uf 2020-05-31 28716.52
#> 39     uf 2020-06-01 28716.52
#> 38     uf 2020-06-02 28716.52
#> 37     uf 2020-06-03 28716.52
#> 36     uf 2020-06-04 28716.52
#> 35     uf 2020-06-05 28716.52
#> 34     uf 2020-06-06 28716.52
#> 33     uf 2020-06-07 28716.52
#> 32     uf 2020-06-08 28716.52
#> 31     uf 2020-06-09 28716.52
#> 30     uf 2020-06-10 28715.56
#> 29     uf 2020-06-11 28714.60
#> 28     uf 2020-06-12 28713.65
#> 27     uf 2020-06-13 28712.69
#> 26     uf 2020-06-14 28711.73
#> 25     uf 2020-06-15 28710.77
#> 24     uf 2020-06-16 28709.82
#> 23     uf 2020-06-17 28708.86
#> 22     uf 2020-06-18 28707.90
#> 21     uf 2020-06-19 28706.94
#> 20     uf 2020-06-20 28705.99
#> 19     uf 2020-06-21 28705.03
#> 18     uf 2020-06-22 28704.07
#> 17     uf 2020-06-23 28703.12
#> 16     uf 2020-06-24 28702.16
#> 15     uf 2020-06-25 28701.20
#> 14     uf 2020-06-26 28700.24
#> 13     uf 2020-06-27 28699.29
#> 12     uf 2020-06-28 28698.33
#> 11     uf 2020-06-29 28697.37
#> 10     uf 2020-06-30 28696.42
#> 9      uf 2020-07-01 28695.46
#> 8      uf 2020-07-02 28694.50
#> 7      uf 2020-07-03 28693.54
#> 6      uf 2020-07-04 28692.59
#> 5      uf 2020-07-05 28691.63
#> 4      uf 2020-07-06 28690.67
#> 3      uf 2020-07-07 28689.72
#> 2      uf 2020-07-08 28688.76
#> 1      uf 2020-07-09 28687.80